+++ src/lib.rs
@@ -111,9 +111,9 @@
 //! license, shall be dual licensed as above, without any additional terms or
 //! conditions.
 //!
-//! [CHANGELOG]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.0/CHANGELOG.md
-//! [LICENSE-MIT]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.0/LICENSE-MIT
-//! [LICENSE-APACHE]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.0/LICENSE-APACHE
+//! [CHANGELOG]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.1/CHANGELOG.md
+//! [LICENSE-MIT]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.1/LICENSE-MIT
+//! [LICENSE-APACHE]: https://github.com/daxpedda/axum-server-dual-protocol/blob/v0.5.1/LICENSE-APACHE
 //! [`axum`]: https://docs.rs/axum/0.6
 //! [`Router`]: https://docs.rs/axum/0.6/axum/struct.Router.html
 //! [`tower`]: https://docs.rs/tower/0.4
@@ -122,6 +122,7 @@ mod dual_protocol;
 mod either;
 mod upgrade_http;
 
+pub use axum_server;
 pub use dual_protocol::{
 	bind_dual_protocol, from_tcp_dual_protocol, DualProtocolAcceptor, DualProtocolAcceptorFuture,
 	DualProtocolService, DualProtocolServiceFuture, Protocol, ServerExt,