+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2023-06-13
+### Added
+- Re-export `axum-server`.
+
 ## [0.5.0] - 2023-06-03
 ### Added
 - Introduced [`Protocol`](https://docs.rs/axum-server-dual-protocol/0.5.0/axum_server_dual_protocol/struct.Protocol.html).
@@ -43,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial commit.
 
+[0.5.1]: https://github.com/daxpedda/axum-server-dual-protocol/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/daxpedda/axum-server-dual-protocol/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/daxpedda/axum-server-dual-protocol/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/daxpedda/axum-server-dual-protocol/compare/v0.2.0...v0.3.0